refactor(front): group all log types into one file
