+++ front/src/features/tables/TableBody.tsx
@@ -1,7 +1,7 @@
-import { llm_logs } from "@prisma/client";
+import { LogDataArray } from "@/types/logDisplayOptions";
 import EmptyLogs from "./EmptyLogs";
 import LogsTable from "./LogsTable";
 
-export const TableBody = ({ logs }: { logs: llm_logs[] }) => {
+export const TableBody = ({ logs }: { logs: LogDataArray }) => {
     return logs.length === 0 ? <EmptyLogs /> : <LogsTable logs={logs} />;
 };
