+++ front/src/hooks/useNavigation.ts
@@ -1,10 +1,10 @@
-import { FilterOptions } from "@/types/filter";
+import { LogDisplayOptions } from "@/types/logDisplayOptions";
 import { useRouter, useSearchParams } from "next/navigation";
 
 export const useNavigation = () => {
     const router = useRouter();
     const searchParams = useSearchParams();
-    const updateSearchParam = (FilterBy: FilterOptions) => {
+    const updateSearchParam = (FilterBy: LogDisplayOptions) => {
         const params = new URLSearchParams(searchParams.toString());
         for (const [key, value] of Object.entries(FilterBy)) {
             params.set(key, value);
