--- front/src/types/sort.ts
@@ -1,2 +1,0 @@
-export type Order = "asc" | "desc";
-export type SortOptions = "datetime_utc" | "total_tokens" | "id";
